+++ examples/etag-hash-cache.ts
@@ -86,7 +86,9 @@ app.route({
 			reply.callNotFound();
 			return;
 		}
-		result.send(reply.raw);
+		await reply.code(result.statusCode)
+			.headers(result.headers)
+			.send(result.stream);
 	},
 });