+++ examples/http-module.ts
@@ -12,7 +12,7 @@ const app = http.createServer((req, res) => {
 			throw new Error('url not set');
 		}
 		const result = await storage.prepareResponse(req.url, req);
-		result.send(res);
+		await result.send(res);
 	})().catch(err => {
 		console.error(err);
 		if (res.headersSent) {