+++ examples/mongodb-gridfs.ts
@@ -103,7 +103,9 @@ client.connect(error => {
 				reply.callNotFound();
 				return;
 			}
-			result.send(reply.raw);
+			await reply.code(result.statusCode)
+				.headers(result.headers)
+				.send(result.stream);
 		},
 	});