+++ examples/pushstate-server.ts
@@ -1,35 +1,36 @@
 
 import { join } from 'path';
 
-import express from 'express';
+import { fastify } from 'fastify';
 
 import { FileSystemStorage, FileSystemStorageError } from '../src/send-stream';
 
-const app = express();
+const app = fastify();
 
 const storage = new FileSystemStorage(join(__dirname, 'assets'));
 
-app.use(async (req, res, next) => {
-	try {
-		let result = await storage.prepareResponse(req.url, req);
+app.route({
+	method: ['HEAD', 'GET'],
+	url: '*',
+	handler: async (request, reply) => {
+		let result = await storage.prepareResponse(request.url, request.raw);
 		// if path is not found then rewrite to root index.html
 		if (result.error instanceof FileSystemStorageError) {
 			result.stream.destroy();
 			const { error: { pathParts } } = result;
 			result = await storage.prepareResponse(
 				['', 'index.html'],
-				req,
+				request.raw,
 				// if the mime type can be determined from path then this is probably an error so add 404 status
-				storage.mimeTypeDefaultCharset(pathParts[pathParts.length - 1])
+				storage.mimeTypeLookup(pathParts[pathParts.length - 1])
 					? { statusCode: 404 }
 					: {},
 			);
 		}
-		result.send(res);
-	} catch (err: unknown) {
-		// eslint-disable-next-line node/callback-return
-		next(err);
-	}
+		await reply.code(result.statusCode)
+			.headers(result.headers)
+			.send(result.stream);
+	},
 });
 
 app.listen(3000, () => {