+++ test/wrappers/server.wrapper.ts
@@ -1,0 +1,29 @@
+import type { Server } from 'http';
+
+import type { Storage, FileSystemStorageOptions, PrepareResponseOptions } from '../../src/send-stream';
+
+export interface ServerWrapper {
+	server?: Server;
+
+	listen(): Promise<void>;
+
+	close(): Promise<void>;
+
+	sendStorage<Reference, AttachedData>(
+		storage: Storage<Reference, AttachedData>,
+		reference: Reference,
+		opts?: PrepareResponseOptions & { noResult?: boolean },
+	): void;
+
+	send(
+		root: string,
+		path?: string | string[],
+		opts?: PrepareResponseOptions & FileSystemStorageOptions & { noResult?: boolean; removeHeader?: string },
+	): void;
+
+	sendWithError(
+		root: string,
+		path?: string | string[],
+		opts?: PrepareResponseOptions & FileSystemStorageOptions & { noResult?: boolean },
+	): void;
+}